Fix bug with client package name not being picked up correctly
Client package name is now sent from client and passed to server correctly
To improve testability of this separate client and service classes setup
Fixed odd test dependency issue with Android studio
Added NoSuchServiceException for bindService failures
Bump to RC2
