--- SampleApp/build.gradle
@@ -1,44 +1,0 @@
-apply plugin: 'com.android.application'
-
-android {
-    compileSdkVersion 28
-
-    defaultConfig {
-        applicationId "com.aevi.android.rxmessenger.sample"
-        minSdkVersion 21
-        targetSdkVersion 28
-        versionCode 1
-        versionName "1.1"
-    }
-
-    buildTypes {
-        release {
-            minifyEnabled false
-            proguardFiles getDefaultProguardFile('proguard-android.txt'), 'proguard-rules.pro'
-        }
-    }
-
-    compileOptions {
-        sourceCompatibility JavaVersion.VERSION_1_8
-        targetCompatibility JavaVersion.VERSION_1_8
-    }
-
-}
-
-dependencies {
-    implementation fileTree(dir: 'libs', include: ['*.jar'])
-
-    implementation project(":rxmessenger")
-
-    implementation presentationDependencies.butterKnife
-    implementation domainDependencies.rxJava
-    implementation domainDependencies.rxAndroid
-    implementation domainDependencies.gson
-    implementation domainDependencies.archLifecycle
-    implementation domainDependencies.archLifecycleJava8
-
-    annotationProcessor annotationDependencies.butterKnife
-
-    implementation(presentationDependencies.support) { force = true }
-    implementation(presentationDependencies.appCompat) { force = true }
-}
