--- SampleApp/src/main/java/com/aevi/android/rxmessenger/sample/MessageTypes.java
@@ -1,9 +1,0 @@
-package com.aevi.android.rxmessenger.sample;
-
-
-public interface MessageTypes {
-
-    String START_ACTIVITY = "startActivity";
-    String RESPONSE = "response";
-    String END_STREAM = "endStream";
-}
