+++ SampleClientApp/proguard-rules.pro
