+++ SampleClientApp/src/main/res/values/colors.xml
