+++ SampleClientApp/src/main/res/values/strings.xml
@@ -1,0 +1,18 @@
+<resources>
+    <string name="app_name">RxMessenger Client Sample</string>
+    <string name="client_activity">RxMessenger Client Activity</string>
+    <string name="sample_activity">Sample Activity</string>
+    <string name="not_connected">Not connected</string>
+    <string name="connect_to_service">Connect to service</string>
+    <string name="send_request">Send request</string>
+    <string name="here_be_response">Here be response</string>
+    <string name="send_response">Send response</string>
+    <string name="send_end">Send End Stream to server</string>
+    <string name="connected_to_service_with_stream">Connected to SampleService (Stream open)</string>
+    <string name="connected_to_service_no_stream">Connected to SampleService (Stream closed)</string>
+    <string name="failed_to_connect">Failed to connect: %s</string>
+    <string name="finish">Finish</string>
+    <string name="received_message">Received message from SampleService: %s</string>
+    <string name="disconnect">Disconnect</string>
+    <string name="received_response">Received response from remote service: %s</string>
+</resources>
