+++ SampleClientApp/src/main/res/values/styles.xml
