+++ build.gradle
@@ -1,4 +1,4 @@
-version = "4.0.0-RC1"
+version = "4.0.0-RC2"
 
 buildscript {
     repositories {
