+++ rxmessenger/src/main/java/com/aevi/android/rxmessenger/client/NoSuchServiceException.java
@@ -1,0 +1,8 @@
+package com.aevi.android.rxmessenger.client;
+
+public class NoSuchServiceException extends RuntimeException {
+
+    public NoSuchServiceException(String message) {
+        super(message);
+    }
+}
