+++ rxmessenger/src/test/java/com/aevi/android/rxmessenger/client/ObservableWebSocketClientTest.java
@@ -5,12 +5,7 @@ import android.content.ComponentName;
 import android.content.Context;
 import android.content.Intent;
 import android.content.pm.ResolveInfo;
-import android.os.Build;
-import android.os.Bundle;
-import android.os.IBinder;
-import android.os.Message;
-import android.os.Messenger;
-import android.os.RemoteException;
+import android.os.*;
 
 import com.aevi.android.rxmessenger.MockShadowMessenger;
 import com.aevi.android.rxmessenger.client.websocket.OkWebSocketClient;
@@ -33,20 +28,13 @@ import java.util.List;
 
 import io.reactivex.observers.TestObserver;
 
-import static com.aevi.android.rxmessenger.MessageConstants.CHANNEL_WEBSOCKET;
-import static com.aevi.android.rxmessenger.MessageConstants.KEY_CHANNEL_TYPE;
-import static com.aevi.android.rxmessenger.MessageConstants.KEY_CLIENT_ID;
-import static com.aevi.android.rxmessenger.MessageConstants.KEY_DATA_RESPONSE;
-import static com.aevi.android.rxmessenger.MessageConstants.MESSAGE_RESPONSE;
+import static com.aevi.android.rxmessenger.MessageConstants.*;
 import static org.assertj.core.api.Java6Assertions.assertThat;
 import static org.mockito.Matchers.anyString;
-import static org.mockito.Mockito.mock;
-import static org.mockito.Mockito.times;
-import static org.mockito.Mockito.verify;
-import static org.mockito.Mockito.when;
+import static org.mockito.Mockito.*;
 import static org.mockito.MockitoAnnotations.initMocks;
 
-@Config(sdk = Build.VERSION_CODES.LOLLIPOP, manifest = Config.NONE, shadows = {MockShadowMessenger.class})
+@Config(manifest = Config.NONE, shadows = {MockShadowMessenger.class})
 @RunWith(RobolectricTestRunner.class)
 public class ObservableWebSocketClientTest {
 
@@ -68,9 +56,11 @@ public class ObservableWebSocketClientTest {
 
     @Test
     public void checkWillHandleNoServiceWithError() throws RemoteException {
+        setupServiceUnbindable();
+
         TestObserver<String> obs = createObservableSendDataAndSubscribe("");
 
-        obs.assertError(RuntimeException.class);
+        obs.assertError(NoSuchServiceException.class);
     }
 
     @Test
@@ -178,6 +168,12 @@ public class ObservableWebSocketClientTest {
     private void verifyServiceIsBound() {
         ShadowApplication shadowApplication = ShadowApplication.getInstance();
         assertThat(shadowApplication.getBoundServiceConnections()).hasSize(1);
+    }
+
+    private void setupServiceUnbindable() {
+        Intent intent = new Intent();
+        intent.setComponent(SERVICE_COMPONENT_NAME);
+        ShadowApplication.getInstance().declareActionUnbindable(intent.getAction());
     }
 
     private void setupMockBoundMessengerService() {
