+++ samplecommonlib/src/main/java/com/aevi/android/rxmessenger/sample/common/MessageTypes.java
@@ -1,0 +1,9 @@
+package com.aevi.android.rxmessenger.sample.common;
+
+
+public interface MessageTypes {
+
+    String START_ACTIVITY = "startActivity";
+    String RESPONSE = "response";
+    String END_STREAM = "endStream";
+}
